make cast to double precision explicit

+++ src/KSPACE/pair_lj_long_tip4p_long.cpp
@@ -1231,7 +1231,7 @@ void PairLJLongTIP4PLong::compute_outer(int eflag, int vflag)
             union_int_float_t t;
             t.f = rsq;
             const int k = (t.i & ncoulmask) >> ncoulshiftbits;
-            double f = (t.f-rtable[k])*drtable[k], qiqj = qtmp*q[j];
+            double f = ((double)t.f-rtable[k])*drtable[k], qiqj = qtmp*q[j];
             if (ni == 0) {
               forcecoul = qiqj*(ftable[k]+f*dftable[k]);
               if (eflag) ecoul = qiqj*(etable[k]+f*detable[k]);
